refactor: update references from UnicornProperties to UnicornApprovals in dependabot configuration and unit test workflows
