+++ unicorn_web/src/schema/unicorn_properties/publicationevaluationcompleted/__init__.py
@@ -2,8 +2,8 @@
 
 from __future__ import absolute_import
 
-from schema.unicorn_properties.publicationevaluationcompleted.marshaller import Marshaller
-from schema.unicorn_properties.publicationevaluationcompleted.AWSEvent import AWSEvent
-from schema.unicorn_properties.publicationevaluationcompleted.PublicationEvaluationCompleted import (
+from schema.unicorn_approvals.publicationevaluationcompleted.marshaller import Marshaller
+from schema.unicorn_approvals.publicationevaluationcompleted.AWSEvent import AWSEvent
+from schema.unicorn_approvals.publicationevaluationcompleted.PublicationEvaluationCompleted import (
     PublicationEvaluationCompleted,
 )
